requirements.txt
# Core data processing
pandas>=2.0.0
numpy>=1.24.0
pyarrow>=13.0.0       # Parquet / Delta table support
openpyxl>=3.1.0

# Data quality & validation
great-expectations>=0.18.0
pydantic>=2.4.0
jsonschema>=4.19.0

# Microsoft Fabric / Azure connectivity
azure-storage-file-datalake>=12.14.0   # ADLS Gen2 (Fabric Lakehouse storage)
azure-identity>=1.15.0                  # DefaultAzureCredential
azure-keyvault-secrets>=4.7.0
notebookutils>=1.0.0; sys_platform == "linux"  # Fabric notebook utils (optional)

# Pipeline orchestration
schedule>=1.2.0
croniter>=2.0.0

# Visualization / UI
streamlit>=1.28.0
plotly>=5.17.0

# Notifications
sendgrid>=6.11.0

# Utilities
python-dotenv>=1.0.0
loguru>=0.7.0
tqdm>=4.66.0
tenacity>=8.2.0
pyyaml>=6.0.1

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
